Make a separate file for global parameters

+++ CS411-Final-Project/Scene_Menu.h
@@ -13,7 +13,7 @@ private:
 	int nButtons;
 	int confirmed;
 
-	int wait;
+	long long wait;
 public:
 	Scene_Menu();
 	~Scene_Menu();
